3DTE: Fixed issue with DHF corrections
The issue was the assumption that all RM measurements would cover the
same time intervals and have the same (or almost the same) beamSeconds
values. If there were any gaps, when the data were resampled to have 30
data points before fitting a model to the data, NaN gaps could arise
where there are no data. Now there is a check for NaN gaps before
fitting the data.